Move GS to gs.ts
