+++ src/devices/gs.ts
@@ -1,12 +1,93 @@
 import {Definition} from '../lib/types';
 import * as exposes from '../lib/exposes';
 import fz from '../converters/fromZigbee';
+import tz from '../converters/toZigbee';
 import * as reporting from '../lib/reporting';
 import {light, onOff, electricityMeter} from '../lib/modernExtend';
 
 const e = exposes.presets;
 
 const definitions: Definition[] = [
+    {
+        zigbeeModel: ['BRHM8E27W70-I1'],
+        model: 'BRHM8E27W70-I1',
+        vendor: 'GS',
+        description: 'Smart color light bulb',
+        extend: [light({colorTemp: {range: undefined}, color: true})],
+    },
+    {
+        zigbeeModel: ['BDHM8E27W70-I1'],
+        model: 'BDHM8E27W70-I1',
+        vendor: 'GS',
+        description: 'Smart light bulb',
+        extend: [light({colorTemp: {range: undefined}})],
+    },
+    {
+        zigbeeModel: ['SGMHM-I1'],
+        model: 'SGMHM-I1',
+        vendor: 'GS',
+        description: 'Methane gas sensor',
+        fromZigbee: [fz.ias_gas_alarm_2],
+        toZigbee: [],
+        exposes: [e.gas()],
+    },
+    {
+        zigbeeModel: ['SGPHM-I1'],
+        model: 'SGPHM-I1',
+        vendor: 'GS',
+        description: 'Propane gas sensor',
+        fromZigbee: [fz.ias_gas_alarm_1],
+        toZigbee: [],
+        exposes: [e.gas(), e.battery_low(), e.tamper()],
+    },
+    {
+        zigbeeModel: ['SKHMP30-I1'],
+        model: 'SKHMP30-I1',
+        vendor: 'GS',
+        description: 'Smart socket',
+        extend: [onOff({powerOnBehavior: false}), electricityMeter()],
+    },
+    {
+        zigbeeModel: ['SMHM-I1', 'PIR_TPV12'],
+        model: 'SMHM-I1',
+        vendor: 'GS',
+        description: 'Motion sensor',
+        fromZigbee: [fz.ias_occupancy_alarm_1, fz.battery],
+        toZigbee: [],
+        exposes: [e.occupancy(), e.battery_low(), e.battery(), e.battery_voltage(), e.tamper()],
+        meta: {battery: {voltageToPercentage: '3V_2500'}},
+        configure: async (device, coordinatorEndpoint, logger) => {
+            const endpoint = device.getEndpoint(1);
+            const bindClusters = ['genPowerCfg'];
+            await reporting.bind(endpoint, coordinatorEndpoint, bindClusters);
+            await reporting.batteryPercentageRemaining(endpoint);
+            await reporting.batteryVoltage(endpoint);
+        },
+    },
+    {
+        zigbeeModel: ['SOHM-I1'],
+        model: 'SOHM-I1',
+        vendor: 'GS',
+        description: 'Open and close sensor',
+        fromZigbee: [fz.ias_contact_alarm_1],
+        toZigbee: [],
+        exposes: [e.contact(), e.battery_low(), e.tamper()],
+    },
+    {
+        zigbeeModel: ['SRHMP-I1'],
+        model: 'SRHMP-I1',
+        vendor: 'GS',
+        description: 'Siren',
+        fromZigbee: [fz.battery, fz.ignore_basic_report],
+        toZigbee: [tz.warning],
+        meta: {disableDefaultResponse: true},
+        configure: async (device, coordinatorEndpoint, logger) => {
+            const endpoint = device.getEndpoint(1);
+            await reporting.bind(endpoint, coordinatorEndpoint, ['genPowerCfg']);
+            await reporting.batteryPercentageRemaining(endpoint);
+        },
+        exposes: [e.battery(), e.warning()],
+    },
     {
         zigbeeModel: ['SSHM-I1'],
         model: 'SSHM-I1',
@@ -22,18 +103,31 @@ const definitions: Definition[] = [
         exposes: [e.smoke(), e.battery_low(), e.tamper(), e.battery()],
     },
     {
-        zigbeeModel: ['BRHM8E27W70-I1'],
-        model: 'BRHM8E27W70-I1',
+        zigbeeModel: ['STHM-I1H'],
+        model: 'STHM-I1H',
         vendor: 'GS',
-        description: 'Smart dimmable, RGB + white (E27 & B22)',
-        extend: [light({colorTemp: {range: undefined}, color: true})],
+        description: 'Temperature and humidity sensor',
+        fromZigbee: [fz.temperature, fz.humidity, fz.battery],
+        toZigbee: [],
+        meta: {battery: {voltageToPercentage: '3V_2500'}},
+        configure: async (device, coordinatorEndpoint, logger) => {
+            const endpoint = device.getEndpoint(1);
+            const bindClusters = ['msTemperatureMeasurement', 'msRelativeHumidity', 'genPowerCfg'];
+            await reporting.bind(endpoint, coordinatorEndpoint, bindClusters);
+            await reporting.temperature(endpoint);
+            await reporting.humidity(endpoint);
+            await reporting.batteryVoltage(endpoint);
+        },
+        exposes: [e.temperature(), e.humidity(), e.battery()],
     },
     {
-        zigbeeModel: ['SKHMP30-I1'],
-        model: 'SKHMP30-I1',
-        description: 'Smart metering plug',
+        zigbeeModel: ['SWHM-I1'],
+        model: 'SWHM-I1',
         vendor: 'GS',
-        extend: [onOff({powerOnBehavior: false}), electricityMeter()],
+        description: 'Water leakage sensor',
+        fromZigbee: [fz.ias_water_leak_alarm_1],
+        toZigbee: [],
+        exposes: [e.water_leak(), e.battery_low(), e.tamper()],
     },
 ];
